Cleaned code - working LQR state reg.
