+++ README.md
@@ -1,7 +1,5 @@
 # 6.8210_project
 
-Yay Ryan has access now!!
-
 Link to the quadrotor example (with an LQR):
 https://deepnote.com/workspace/Underactuated-2ed1518a-973b-4145-bd62-1768b49956a8/project/03-Acrobots-Cart-Poles-and-Quadrotors-096cffe7-416e-4d51-a471-5fd526ec8fab/notebook/acrobot-229d29e61b5b44cabccbc7ac3dd59881
